// src/models/produto.ts
// Modelos de produto
export interface Produto {
    id: number;
    unidadeProdutoId?: number;
    gtin?: string;
    codigoInterno?: string;
    nome: string;
    descricao?: string;
    descricaoPdv?: string;
    valorVenda?: number;
    quantidadeEstoque?: number;
    quantidadeEstoqueAnterior?: number;
    estoqueMin?: number;
    estoqueMax?: number;
    iat?: string;
    ippt?: string;
    ncm?: string;
    tipoItemSped?: string;
    dataEstoque?: Date;
    horaEstoque?: string;
    taxaIpi?: number;
    taxaIssqn?: number;
    taxaPis?: number;
    taxaCofins?: number;
    taxaIcms?: number;
    cst?: string;
    csosn?: string;
    totalizadorParcial?: string;
    ecfIcmsSt?: string;
    codigoBalanca?: number;
    pafPSt?: string;
    hashTripa?: string;
    hashIncremento?: string;
    desativado?: boolean;
    cozinha?: boolean;
    categoriaProdutoId?: number;
    lucro?: number;
    valorCompra?: number;
    grupoId?: number;
    subgrupoId?: number;
    valorPromocao?: number;
    servico?: boolean;
    fornecedorId?: number;
    cfop?: number;
    atacadoAcimade?: number;
    atacadoValorVenda?: number;
    cest?: string;
    dataVencimento?: Date;
    cstIpi?: string;
    cstCofins?: string;
    cfopSaida?: number;
    quantidadeEstoqueSemNota?: number;
    foto:string;
  }  